2023/6/28
1.完善了目标检测函数
2.完善了卡片识别函数，减小识别的范围
3.完善了OPENART2的目标检测函数

+++ Project/CODE/openart2.h
@@ -3,8 +3,8 @@
 
 #include "headfile.h"
 
-extern int16 ART2_center_x;
-extern int16 ART2_center_y;
+extern int8 ART2_center_x;
+extern int8 ART2_center_y;
 extern int8 ART2_angle;
 extern uint8 ART2_mode;
